backend/src/upload_trigger/requirements.txt
boto3==1.28.57
botocore==1.31.57
PyPDF2==3.0.1
shortuuid==1.0.11
python-docx
